Add basic orbit functionality

+++ src/core/window_impl.cpp
@@ -127,8 +127,8 @@ static auto glfw_cursor_pos_callback(GLFWwindow* window, double x, double y) -> 
 
     event->type = MouseEvent::Type::Moved;
     event->button = MouseButton::None;
-    event->position = {x, y};
-    event->scroll = {0.0, 0.0};
+    event->position = {static_cast<float>(x), static_cast<float>(y)};
+    event->scroll = {0.0f, 0.0f};
 
     EventDispatcher::Get().Dispatch("mouse_event", std::move(event));
 }
@@ -139,8 +139,11 @@ static auto glfw_mouse_button_callback(GLFWwindow* window, int button, int actio
 
     event->type = MouseEvent::Type::ButtonPressed;
     event->button = glfw_mouse_button_map(button);
-    event->position = {instance->mouse_pos_x, instance->mouse_pos_y};
-    event->scroll = {0.0, 0.0};
+    event->position = {
+        static_cast<float>(instance->mouse_pos_x),
+        static_cast<float>(instance->mouse_pos_y)
+    };
+    event->scroll = {0.0f, 0.0f};
 
     if (action == GLFW_PRESS) {
         EventDispatcher::Get().Dispatch("mouse_event", std::move(event));
@@ -158,8 +161,11 @@ static auto glfw_scroll_callback(GLFWwindow* window, double x, double y) -> void
 
     event->type = MouseEvent::Type::Scrolled;
     event->button = MouseButton::None;
-    event->position = {instance->mouse_pos_x, instance->mouse_pos_y};
-    event->scroll = {x, y};
+    event->position = {
+        static_cast<float>(instance->mouse_pos_x),
+        static_cast<float>(instance->mouse_pos_y)
+    };
+    event->scroll = {static_cast<float>(x), static_cast<float>(y)};
 
     EventDispatcher::Get().Dispatch("mouse_event", std::move(event));
 }
